requirements.txt
python-dotenv
requests
groq
PyMuPDF